Refactor Headers into headers.rs
